+++ CHANGELOG.md
@@ -3,6 +3,18 @@ Polyverse Boost ReST API (Backend)
 
 # Release Notes
 
+## Version 0.13.1: March 11th, 2024
+
+### New Features
+- Added REST API to delete project grooming data (DELETE) - /api/user_project/{org}/{project}/groom
+    - Automatically triggered if a grooming request (POST) is made against a deleted project
+
+### Enhancements
+- Improved logging for Project Grooming and Discovery Grooming - to include more details on the grooming process
+
+### Bug Fixes
+- N/A
+
 ## Version 0.13.0: March 8th, 2024
 
 ### New Features